Enum ata::AtaDeviceType

pub enum AtaDeviceType {
    Pata,
    PataPi,
    Sata,
    SataPi,
}

The possible types of drive devices that can be attached to an IDE controller via ATA.

Variants

Pata

A parallel ATA (PATA) drive, like a hard drive. This is the type previously known as just “ATA” before SATA existed.

PataPi

A parallel ATA (PATA) drive that uses the packet interface, like an optical CD-ROM drive.

Sata

A serial ATA (SATA) drive that is operating in legacy IDE emulation mode, not the standard AHCI interface for SATA. Some systems refer to this as a SEMB (SATA Enclosure Management Bridge) device, which may or may not be attached through a port multiplier.

SataPi

A serial ATA (SATA) drive that that is operating in legacy IDE emulation mode and uses the packet interface.