Struct ata::AtaDrive

pub struct AtaDrive { /* fields omitted */ }

A single ATA drive, either a master or a slave, which sits on one of two buses within a larger IDE controller.

Implementations

impl AtaDrive

pub fn read_pio(
&mut self,
buffer: &mut [u8],
offset_in_sectors: usize
) -> Result<usize, &'static str>

Reads data from this drive starting at the given offset_in_sectors into the provided buffer. The length of the given buffer determines the number of bytes to be written.

As content is read from the drive at sector granularity, the buffer length must be a multiple of the sector size (512 bytes), and the offset is specified in number of sectors (not number of bytes) from the beginning of the drive.

Returns the number of sectors (not bytes) that were successfully written to the drive.

Note

This is slow, as it uses blocking port I/O instead of DMA.

pub fn write_pio(
&mut self,
buffer: &[u8],
offset_in_sectors: usize
) -> Result<usize, &'static str>

Writes data from the provided buffer to this drive, starting at the given offset_in_sectors into the drive. The length of the given buffer determines the number of bytes to be written.

As content is written to the drive at sector granularity, the buffer length must be a multiple of the sector size (512 bytes), and the offset is specified in number of sectors (not number of bytes) from the beginning of the drive.

Returns the number of sectors (not bytes) that were successfully written to the drive.

Note

This is slow, as it uses blocking port I/O instead of DMA.

pub fn is_master(&self) -> bool

Returns true if this drive is the master, or false if it is the slave on the IDE controller bus.