Struct ata::AtaIdentifyData

#[repr(packed)]
pub struct AtaIdentifyData {

    pub general_configuration: u16,
    pub num_cylinders: u16,
    pub specific_configuration: u16,
    pub num_heads: u16,
    pub num_sectors_per_track: u16,
    pub vendor_unique1: [u16; 3],
    pub serial_number: AtaSerialNumber,
    pub firmware_version: AtaFirmwareVersion,
    pub model_number: AtaModelNumber,
    pub max_blocks_per_transfer: u8,
    pub vendor_unique2: u8,
    pub trusted_computing: u16,
    pub capabilities: u16,
    pub translation_fields_valid: u8,
    pub free_fall_control_sensitivity: u8,
    pub num_current_cylinders: u16,
    pub num_current_heads: u16,
    pub current_sectors_per_track: u16,
    pub current_sector_capacity: u32,
    pub current_multi_sector_setting: u8,
    pub ext_command_supported: u8,
    pub user_addressable_sectors: u32,
    pub multiword_dma_support: u8,
    pub multiword_dma_active: u8,
    pub advanced_pio_modes: u8,
    pub minimum_mw_transfer_cycle_time: u16,
    pub recommended_mw_transfer_cycle_time: u16,
    pub minimum_pio_cycle_time: u16,
    pub minimum_pio_cycle_time_io_ready: u16,
    pub additional_supported: u16,
    pub queue_depth: u16,
    pub serial_ata_capabilities: u32,
    pub serial_ata_features_supported: u16,
    pub serial_ata_features_enabled: u16,
    pub major_revision: u16,
    pub minor_revision: u16,
    pub command_set_support: [u16; 3],
    pub command_set_active: [u16; 3],
    pub ultra_dma_support: u8,
    pub ultra_dma_active: u8,
    pub normal_security_erase_unit: u16,
    pub enhanced_security_erase_unit: u16,
    pub current_apm_level: u8,
    pub master_password_id: u16,
    pub hardware_reset_result: u16,
    pub current_acoustic_value: u8,
    pub recommended_acoustic_value: u8,
    pub stream_min_request_size: u16,
    pub streaming_transfer_time_dma: u16,
    pub streaming_access_latency_dma_pio: u16,
    pub streaming_perf_granularity: u32,
    pub max_48_bit_lba: u64,
    pub streaming_transfer_time: u16,
    pub dsm_cap: u16,
    pub physical_logical_sector_size: u16,
    pub inter_seek_delay: u16,
    pub world_wide_name: [u16; 4],
    pub reserved_for_world_wide_name_128: [u16; 4],
    pub reserved_for_tlc_technical_report: u16,
    pub words_per_logical_sector: u32,
    pub command_set_support_ext: u16,
    pub command_set_active_ext: u16,
    pub reserved_for_expanded_support_and_active: [u16; 6],
    pub msn_support: u16,
    pub security_status: u16,
    pub cfa_power_mode1: u16,
    pub nominal_form_factor: u16,
    pub data_set_management_feature: u16,
    pub additional_product_id: [u16; 4],
    pub current_media_serial_number: [u16; 30],
    pub sct_command_transport: u16,
    pub block_alignment: u16,
    pub write_read_verify_sector_count_mode_3_only: [u16; 2],
    pub write_read_verify_sector_count_mode_2_only: [u16; 2],
    pub nv_cache_capabilities: u16,
    pub nv_cache_size_lsw: u16,
    pub nv_cache_size_msw: u16,
    pub nominal_media_rotation_rate: u16,
    pub nv_cache_time_to_spin_up_in_seconds: u8,
    pub write_read_verify_sector_count_mode: u8,
    pub transport_major_version: u16,
    pub transport_minor_version: u16,
    pub extended_num_of_user_addressable_sectors: u64,
    pub min_blocks_per_download_microcode: u16,
    pub max_blocks_per_download_microcode: u16,
    pub signature: u8,
    pub checksum: u8,
    // some fields omitted

}

Information that describes an ATA drive, obtained from the response to an identify command.

Fuller documentation is available here: https://docs.microsoft.com/en-us/windows-hardware/drivers/ddi/content/ata/ns-ata-_identify_device_data#members

Fields
general_configuration: u16num_cylinders: u16specific_configuration: u16num_heads: u16num_sectors_per_track: u16vendor_unique1: [u16; 3]serial_number: AtaSerialNumberfirmware_version: AtaFirmwareVersionmodel_number: AtaModelNumbermax_blocks_per_transfer: u8

Maximum number of blocks per transfer. Sometimes referred to as “sectors per int”.

vendor_unique2: u8trusted_computing: u16capabilities: u16translation_fields_valid: u8

A bitmask of translation fields valid and free fall control sensitivity

free_fall_control_sensitivity: u8num_current_cylinders: u16num_current_heads: u16current_sectors_per_track: u16current_sector_capacity: u32current_multi_sector_setting: u8ext_command_supported: u8

MultiSectorSettingValid : 1; ReservedByte59 : 3; SanitizeFeatureSupported : 1; CryptoScrambleExtCommandSupported : 1; OverwriteExtCommandSupported : 1; BlockEraseExtCommandSupported : 1;

user_addressable_sectors: u32

Number of sectors in the drive, if using 28-bit LBA. This can be used to calculate the size of the drive. If zero, we’re using 48-bit LBA, so you should use max_48_bit_lba.

multiword_dma_support: u8multiword_dma_active: u8advanced_pio_modes: u8minimum_mw_transfer_cycle_time: u16recommended_mw_transfer_cycle_time: u16minimum_pio_cycle_time: u16minimum_pio_cycle_time_io_ready: u16additional_supported: u16queue_depth: u16

only the first 5 bits are used, others are reserved

serial_ata_capabilities: u32serial_ata_features_supported: u16serial_ata_features_enabled: u16major_revision: u16minor_revision: u16command_set_support: [u16; 3]command_set_active: [u16; 3]ultra_dma_support: u8ultra_dma_active: u8normal_security_erase_unit: u16enhanced_security_erase_unit: u16current_apm_level: u8master_password_id: u16hardware_reset_result: u16current_acoustic_value: u8recommended_acoustic_value: u8stream_min_request_size: u16streaming_transfer_time_dma: u16streaming_access_latency_dma_pio: u16streaming_perf_granularity: u32max_48_bit_lba: u64

Number of sectors in the drive, if using 48-bit LBA. This can be used to calculate the size of the drive.

streaming_transfer_time: u16dsm_cap: u16physical_logical_sector_size: u16

[0:3] Physical sector size (in logical sectors)

inter_seek_delay: u16world_wide_name: [u16; 4]reserved_for_world_wide_name_128: [u16; 4]reserved_for_tlc_technical_report: u16words_per_logical_sector: u32command_set_support_ext: u16command_set_active_ext: u16reserved_for_expanded_support_and_active: [u16; 6]msn_support: u16security_status: u16cfa_power_mode1: u16nominal_form_factor: u16data_set_management_feature: u16additional_product_id: [u16; 4]current_media_serial_number: [u16; 30]sct_command_transport: u16block_alignment: u16write_read_verify_sector_count_mode_3_only: [u16; 2]write_read_verify_sector_count_mode_2_only: [u16; 2]nv_cache_capabilities: u16nv_cache_size_lsw: u16nv_cache_size_msw: u16nominal_media_rotation_rate: u16nv_cache_time_to_spin_up_in_seconds: u8write_read_verify_sector_count_mode: u8transport_major_version: u16transport_minor_version: u16extended_num_of_user_addressable_sectors: u64min_blocks_per_download_microcode: u16max_blocks_per_download_microcode: u16signature: u8checksum: u8