Struct ata::AtaSerialNumber

#[repr(packed)]
pub struct AtaSerialNumber(_);

An ATA drive’s serial number is a 20-byte string.

This is a wrapper around a byte string [u8; 20], because Rust only supports deriving traits like Debug and Default for arrays up to 32 elements.