Struct ata::IdeController

pub struct IdeController {
    pub primary_master: Option<AtaDriveRef>,
    pub primary_slave: Option<AtaDriveRef>,
    pub secondary_master: Option<AtaDriveRef>,
    pub secondary_slave: Option<AtaDriveRef>,
}

A single IDE controller has two buses with up to two drives attached to each bus, for a total of up to four drives.

Fields
primary_master: Option<AtaDriveRef>primary_slave: Option<AtaDriveRef>secondary_master: Option<AtaDriveRef>secondary_slave: Option<AtaDriveRef>
Implementations

impl IdeController

pub fn new(pci_device: &PciDevice) -> Result<IdeController, &'static str>

Creates a new instance of an IDE controller based on the given PCI device.

pub fn iter(&self) -> IdeControllerIter<'_>

Returns an Iterator over all of the AtaDrives that exist (and are supported) in this IdeController. The order of iteration is: primary master, primary slave, secondary master, and secondary slave; any devices that do not exist are skipped.