Crate ata

Support for accessing ATA drives (IDE).

The primary struct of interest is AtaDrive.

Structs

AtaDrive
A single ATA drive, either a master or a slave, which sits on one of two buses within a larger IDE controller.

AtaError
The possible error values found in an ATA drive’s error port.

AtaFirmwareVersion
An ATA drive’s firmware version is an 8-byte string.

AtaIdentifyData
Information that describes an ATA drive, obtained from the response to an identify command.

AtaModelNumber
An ATA drive’s model number is a 40-byte string.

AtaSerialNumber
An ATA drive’s serial number is a 20-byte string.

AtaStatus
The possible status values found in an ATA drive’s status port.

IdeController
A single IDE controller has two buses with up to two drives attached to each bus, for a total of up to four drives.

IdeControllerIter
Provides an iterator over all AtaDrives in an IdeController. See the IdeController::iter() method.

Enums

AtaDeviceType
The possible types of drive devices that can be attached to an IDE controller via ATA.

Type Definitions

AtaDriveRef